+++ examples/asset_loading/main.rs
@@ -125,10 +125,9 @@ fn main() -> Result<(), Error> {
     let app_root = application_root_dir()?;
 
     // Add our meshes directory to the asset loader.
-    let resources_directory = app_root.join("examples/assets");
+    let assets_directory = app_root.join("examples/assets");
 
-    let display_config_path =
-        app_root.join("{}/examples/asset_loading/resources/display_config.ron");
+    let display_config_path = app_root.join("{}/examples/asset_loading/config/display.ron");
 
     let game_data = GameDataBuilder::default()
         .with_bundle(WindowBundle::from_config_path(display_config_path))?
@@ -137,7 +136,7 @@ fn main() -> Result<(), Error> {
         .with_thread_local(RenderingSystem::<DefaultBackend, _>::new(
             ExampleGraph::default(),
         ));
-    let mut game = Application::new(resources_directory, AssetsExample, game_data)?;
+    let mut game = Application::new(assets_directory, AssetsExample, game_data)?;
     game.run();
     Ok(())
 }