+++ examples/auto_fov/main.rs
@@ -51,8 +51,8 @@ fn main() -> Result<(), Error> {
     amethyst::start_logger(Default::default());
 
     let app_dir = amethyst::utils::application_dir("examples")?;
-    let display_config_path = app_dir.join("auto_fov/resources/display.ron");
-    let assets = app_dir.join("assets");
+    let display_config_path = app_dir.join("auto_fov/config/display.ron");
+    let assets_directory = app_dir.join("assets");
 
     let game_data = GameDataBuilder::new()
         .with_bundle(WindowBundle::from_config_path(display_config_path))?
@@ -66,7 +66,7 @@ fn main() -> Result<(), Error> {
             ExampleGraph::default(),
         ));
 
-    let mut game = Application::build(assets, Loading::new())?.build(game_data)?;
+    let mut game = Application::build(assets_directory, Loading::new())?.build(game_data)?;
     game.run();
 
     Ok(())