+++ examples/sprite_animation/main.rs
@@ -145,8 +145,7 @@ fn main() -> amethyst::Result<()> {
 
     let app_root = application_root_dir()?;
     let assets_directory = app_root.join("examples/assets/");
-    let display_config_path =
-        app_root.join("examples/sprite_animation/resources/display_config.ron");
+    let display_config_path = app_root.join("examples/sprite_animation/config/display.ron");
 
     let game_data = GameDataBuilder::default()
         .with_bundle(WindowBundle::from_config_path(display_config_path))?